+++ expenses/views/expenses.py
@@ -3,14 +3,15 @@ import io
 import requests
 from datetime import date
 
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.db.models import Q, Sum
 from django.conf import settings
 from django.forms import ValidationError
 from django.urls import reverse_lazy, reverse
 from django.utils import timezone
-from django.views.generic import FormView, ListView
+from django.views.generic import CreateView, DeleteView, FormView, ListView, UpdateView
 
-from expenses.forms import ExpenseFileUploadForm
+from expenses.forms import ExpenseFileUploadForm, ExpenseForm
 from expenses.models import (
     Account,
     Currency,
@@ -46,67 +47,28 @@ class UploadExpenseView(FormView):
         decoded_file = io.TextIOWrapper(file, encoding="utf-8-sig", newline="")
         csv_reader = csv.reader(decoded_file)
 
-        try:
-            default_currency = Currency.objects.get(alpha3=settings.DEFAULT_CURRENCY)
-        except Currency.DoesNotExist:
-            raise ValidationError("Default currency not found")
-
-        try:
-            default_account = Account.objects.get(name=settings.DEFAULT_ACCOUNT)
-        except Account.DoesNotExist:
-            raise ValidationError("Default account not found")
-
         lines = 0
         for row in csv_reader:
             lines += 1
             print(f"Line: {lines} - {row}")
             row = self._clear_row(row)
-            q_account = Account.objects.filter(name=row[3])
 
-            if q_account.exists():
-                account = q_account.first()
-            else:
-                account = default_account
-                print("Default account")
+            payment_date = self._get_payment_date(row[0])
+            period = Period.get_period_from_date(payment_date)
 
-            money = row[2].split(" ")
-            try:
-                amount = float(money[0].replace(",", ""))
-            except (IndexError, ValueError) as e:
-                print("Error: amount not found")
-                continue
-
-            try:
-                code_currency = money[1]
-            except IndexError as e:
-                code_currency = None
-
-            currency = Currency.objects.filter(alpha3=code_currency)
-            if not currency.exists():
-                print("Warning: using default currency")
-                currency = default_currency
-            else:
-                currency = currency.first()
-
-            payment_date = str_to_date(row[0])
-            try:
-                period = Period.objects.get(
-                    year=payment_date.year, month=payment_date.month
-                )
-            except Period.DoesNotExist:
-                print("Error: Period does not exists")
+            if not period:
+                print("Period not set")
                 continue
 
             if period.closed:
                 print("Error: Period is closed")
                 continue
 
-            if not payment_date:
-                payment_date = timezone.now().date()
+            amount, currency = self._get_amount(row[2])
+            account = self._get_account(row[3])
 
             if Expense.objects.filter(
                 period=period,
-                account=account,
                 currency=currency,
                 description=row[1],
                 amount=amount,
@@ -125,9 +87,7 @@ class UploadExpenseView(FormView):
                 }
             )
             if serializer.is_valid():
-                expense = serializer.save()
-                expense.local_amount = expense.get_local_amount
-                expense.save()
+                serializer.save()
                 print("Expense saved!!")
             else:
                 print(serializer.errors)
@@ -148,6 +108,55 @@ class UploadExpenseView(FormView):
             print("POST request sent successfully")
         else:
             print("Error: Failed to send POST request")
+
+    def _get_account(self, value: str):
+        account_q = Account.objects.filter(name=value)
+
+        if account_q.exists():
+            account = account_q.first()
+        else:
+            account = Account.get_default
+            print("Default account")
+
+        return account
+
+    def _get_amount(self, value: str) -> tuple:
+        """
+        Asumming that the value is: [number currency]
+        """
+        money = value.split(" ")
+        try:
+            amount = float(money[0].replace(",", ""))
+        except (IndexError, ValueError) as e:
+            amount = 0
+
+        try:
+            code_currency = money[1]
+        except IndexError:
+            code_currency = None
+
+        currency_q = Currency.objects.filter(alpha3=code_currency)
+        if currency_q.exists():
+            currency = currency_q.first()
+        else:
+            currency = Currency.get_default
+
+        return (amount, currency)
+
+    def _get_period(self, value: date):
+        period = Period.get_period_from_date(value)
+
+        if period.closed:
+            print("Error: Period is closed")
+            continue
+
+    def _get_payment_date(self, value: str):
+        payment_date = str_to_date(value)
+
+        if not payment_date:
+            payment_date = timezone.now().date()
+
+        return payment_date
 
 
 class ExpenseGroupListView(ListView):
@@ -179,9 +188,9 @@ class ExpenseGroupListView(ListView):
         return context
 
 
-class ExpenseListView(ListView):
+class ExpenseListPerAccountView(ListView):
     model = Expense
-    template_name = "expenses/expense_list.html"
+    template_name = "expenses/expense_per_account_list.html"
     context_object_name = "expenses"
 
     def get_queryset(self):
@@ -203,3 +212,48 @@ class ExpenseListView(ListView):
         context["account"] = account.name
         context["total"] = get_total_local_amount(Q(period=period, account=account))
         return context
+
+
+class ExpenseListView(ListView):
+    model = Expense
+    template_name = "expenses/expense_list.html"
+    context_object_name = "expenses"
+    paginate_by = 12
+    ordering = ["-period", "-payment_date"]
+
+    def get_context_data(self, **kwargs):
+        context = super(ExpenseListView, self).get_context_data(**kwargs)
+        expenses = Expense.objects.all()
+        paginator = Paginator(expenses, self.paginate_by)
+
+        page = self.request.GET.get("page")
+
+        try:
+            expenses = paginator.page(page)
+        except PageNotAnInteger:
+            expenses = paginator.page(1)
+        except EmptyPage:
+            expenses = paginator.page(paginator.num_pages)
+
+        context["expenses"] = expenses
+        return context
+
+
+class ExpenseCreateView(CreateView):
+    model = Expense
+    form_class = ExpenseForm
+    template_name = "expenses/expense_form.html"
+    success_url = reverse_lazy("expense-list")
+
+
+class ExpenseUpdateView(UpdateView):
+    model = Expense
+    form_class = ExpenseForm
+    template_name = "expenses/expense_form.html"
+    success_url = reverse_lazy("expense-list")
+
+
+class ExpenseDeleteView(DeleteView):
+    model = Expense
+    template_name = "expenses/expense_confirm_delete.html"
+    success_url = reverse_lazy("expense-list")